+++ tests/testthat/test-connector.R
@@ -38,3 +38,20 @@ cli::test_that_cli("Test connector creation", {
     expect_snapshot_out(print(connector_obj$test))
   })
 })
+
+cli::test_that_cli("can create Connector object", {
+
+  temp_dir <- withr::local_tempdir("connector_test")
+
+  withr::with_tempdir(tmpdir = temp_dir, {
+
+    test <- ConnectorFS$new(path = ".")
+
+    connector_obj <- connectors(
+      test = test,
+      test_2 = base::as.data.frame(x = iris)
+    )
+
+    expect_snapshot_out(print(datasources(connector_obj)))
+  })
+})